fix: call .bind and move test file
